CHANGE first-pass at removing all PHP warnings and JS console errors when WP_DEBUG is turned on. Includes fixes for AJAX handlers ADD filter to determine which 'empty' tags are allowed in fields. 'profile_cct_strip_tags'

+++ views/page.php
@@ -1,6 +1,9 @@
 <?php
 	if ( isset( $_REQUEST['reset'] ) && isset( $_REQUEST['nonce'] ) && wp_verify_nonce( $_REQUEST['nonce'], 'profile_cct_reset_fields' ) ):
-		Profile_CCT_Admin::delete_option( 'page', 'tabs', 'normal', $tabs ); // Reset to defaults.
+		if( isset( $tabs ) ){
+			Profile_CCT_Admin::delete_option( 'page', 'tabs', 'normal', $tabs ); // Reset to defaults.	
+		}
+		
 		foreach( self::get_contexts() as $context ):
 			Profile_CCT_Admin::delete_option( 'page', 'fields', $context );
 		endforeach;
